Execution/requirements.txt
accelerate==0.22.0
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
coloredlogs==15.0.1
datasets==2.14.4
dill==0.3.7
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.9.0
google-generativeai==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
lit==16.0.6
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
nltk==3.8.1
numpy==1.24.4
omegaconf==2.3.0
openai==0.28.0
optimum==1.8.8
packaging==23.1
pandas==2.0.3
peft==0.5.0
Pillow==10.0.0
protobuf==3.20.0
psutil==5.9.5
pyarrow==13.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.29.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2